requirements.txt
asgiref==3.8.1
attrs==24.2.0
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.4.0
Django==5.1.3
django-appconf==1.0.6
django-cors-headers==4.6.0
django-imagekit==5.0.0
djangorestframework==3.15.2
drf-spectacular==0.27.2
drf-spectacular-sidecar==2024.11.1
idna==3.10
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
pilkit==3.0
pillow==11.0.0
psycopg2-binary==2.9.10
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rpds-py==0.21.0
soupsieve==2.6
sqlparse==0.5.1
uritemplate==4.1.1
urllib3==2.2.3
